optee: fix machine suffix handling and add tee.srec symlink
The previous install rule hardcoded the "spider" suffix in the
"tee.srec" filename, regardless of the selected machine
("spider" or "s4sk").

This caused an error when running the "boot_artifacts" for the "s4sk"
machine, because it expected "tee-s4sk.srec", but really
"tee-spider.srec" was created.

Now the machine-suffix is displayed correctly.
Additionally, a symlink link "tee.srec" has been added, which will
always point to the correct file regardless of the selected machine.

Signed-off-by: Mykhailo Androsiuk <mykhailo_androsiuk@epam.com>